demo - day 4

+++ app/Http/Controllers/DepartmentController.php
@@ -4,11 +4,22 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Department;
+use Illuminate\Support\Facades\Gate; // Import the Gate facade
 class DepartmentController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+
+     // check the user is authenticated or not
+    public function __construct(){
+        // $this->middleware('auth'); // all controller methods are protected by auth middleware
+        // $this->middleware('auth')->except(['index', 'show']); // only index method is public, all other methods are protected by auth middleware
+        $this->middleware('auth')->only(['index']); // only index method is protected by auth middleware
+    }
+
+
+
     public function index()
     {
         $departments = Department::all();
@@ -28,8 +39,18 @@ class DepartmentController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
-        $department = Department::create($request->all()); // don't forget to add the fillable attributes in the model 
+    {   
+
+        $department = new Department();
+        $department->name = $request->name;
+        $department->description = $request->description;
+        $department->user_id = auth()->user()->id; // catch all data of the logged in user 
+        $department->save();
+
+        // dd(auth()->user()->id);
+        // $department = Department::create($request->all()); // don't forget to add the fillable attributes in the model 
+            // $department->user_id = auth()->user()->id; // set the user_id to the authenticated user id
+        
         // rfedirect to index page
         return to_route('departments.index');
         // ->with('success', 'Department created successfully');
@@ -72,7 +93,14 @@ class DepartmentController extends Controller
      */
     public function destroy(string $id)
     {
-        $department = Department::findorfail($id);
+        $department = Department::findorfail($id); // id
+         // get the authenticated user
+
+        if(Gate::allows("manage-department", $department) === false){
+            // if the user is not allowed to delete the department, redirect to index page with error message
+            return to_route('departments.index');
+        }
+
         $department->delete();
         // redirect to index page
         return to_route('departments.index');
